refactor: 모임 수정 모달 리팩토링 및 일부 스타일링 수정
