+++ frontend/src/components/GroupEdit/index.tsx
@@ -1,42 +1,33 @@
-import { useEffect, useState } from 'react';
-
 import { useQueryClient } from 'react-query';
+import { useLocation } from 'react-router-dom';
 import { useRecoilValue, useResetRecoilState } from 'recoil';
 
 import { requestEditGroup } from 'apis/request/group';
 import Modal from 'components/Modal';
 import { QUERY_KEY } from 'constants/key';
 import { GUIDE_MESSAGE } from 'constants/message';
-import { GROUP_RULE } from 'constants/rule';
-import useClosingState from 'hooks/useClosingState';
 import useCreateState from 'hooks/useCreateState';
 import useModal from 'hooks/useModal';
 import useSnackbar from 'hooks/useSnackbar';
-import CalendarEditor from 'components/CalendarEditor';
 import validateGroupData from 'pages/Create/validate';
-import { categoryState, groupDetailState, modalState } from 'store/states';
-import { CategoryType } from 'types/data';
-import { getNewDateString } from 'utils/date';
+import { groupDetailState, modalState } from 'store/states';
 import PageError from 'utils/PageError';
+
+import Inputs from './Inputs';
 
 import * as S from './index.styled';
 
-const dropdownAnimationTime = 300;
-
 function GroupEdit() {
   const modalFlag = useRecoilValue(modalState);
-  const categories = useRecoilValue(categoryState);
-  const groupData = useRecoilValue(groupDetailState);
-  const resetGroupData = useResetRecoilState(groupDetailState);
 
   const { setOffModal } = useModal();
   const { setMessage } = useSnackbar();
+
   const queryClient = useQueryClient();
-  const [isShownDropdown, setIsShownDropdown] = useState(false);
-  const { isClosing, close } = useClosingState(dropdownAnimationTime, () => {
-    setIsShownDropdown(false);
-  });
-  const params = window.location.pathname.split('/');
+
+  const resetGroupData = useResetRecoilState(groupDetailState);
+
+  const params = useLocation().pathname.split('/');
   const id = params[2];
 
   const {
@@ -50,66 +41,8 @@ function GroupEdit() {
     useDescriptionState,
     getGroupState,
   } = useCreateState();
-  const duration = {
-    start: getGroupState().startDate,
-    end: getGroupState().endDate,
-  };
 
-  const { name, setName, dangerouslySetName } = useNameState();
-  const { selectedCategory, setSelectedCategory } = useSelectedCategoryState();
-  const { capacity, setCapacity } = useCapacityState();
-  const {
-    startDate,
-    setStartDate,
-    endDate,
-    setEndDate,
-    dangerouslySetStartDate,
-    dangerouslySetEndDate,
-  } = useDateState();
-  const { setSchedules } = useScheduleState();
-  const { deadline, setDeadline, dangerouslySetDeadline } = useDeadlineState();
-  const { location, setLocation, dangerouslySetLocation } = useLocationState();
-  const { description, setDescription, dangerouslySetDescription } =
-    useDescriptionState();
-
-  useEffect(() => {
-    if (groupData.id === -1) return;
-
-    dangerouslySetName(groupData.name);
-    setSelectedCategory({
-      id: groupData.categoryId,
-      name: categories[groupData.categoryId - 1 || 0].name,
-    });
-    setCapacity(groupData.capacity);
-    dangerouslySetStartDate(groupData.duration.start);
-    dangerouslySetEndDate(groupData.duration.end);
-    groupData.schedules.forEach(schedule => setSchedules(schedule));
-    dangerouslySetDeadline(groupData.deadline);
-    dangerouslySetLocation(groupData.location);
-    dangerouslySetDescription(groupData.description);
-  }, [modalFlag]);
-
-  if (!groupData) return <></>;
-
-  const toggleDropdownState = () => {
-    if (isShownDropdown) {
-      close();
-      return;
-    }
-
-    setIsShownDropdown(true);
-  };
-
-  const changeCategory = (item: CategoryType) => () => {
-    setSelectedCategory(item);
-    close();
-  };
-
-  const changeCapacity = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setCapacity(Number(e.target.value));
-  };
-
-  const groupEdit = () => {
+  const editGroup = () => {
     const groupData = getGroupState();
 
     try {
@@ -126,7 +59,9 @@ function GroupEdit() {
       .then(() => {
         setOffModal();
         resetGroupData();
+
         queryClient.invalidateQueries(QUERY_KEY.GROUP_DETAILS);
+
         setMessage(GUIDE_MESSAGE.GROUP.SUCCESS_EDIT_REQUEST);
       })
       .catch(error => {
@@ -138,104 +73,20 @@ function GroupEdit() {
     <Modal modalState={modalFlag === 'groupEdit'}>
       <S.Container>
         <S.Title>모임 수정</S.Title>
-        <S.InputContainer>
-          {groupData && (
-            <>
-              <S.Label>
-                이름
-                <S.Input
-                  type="text"
-                  placeholder="정체를 밝혀라 @-@"
-                  value={name}
-                  onChange={setName}
-                />
-              </S.Label>
-              <S.Label>
-                카테고리
-                <S.CategoryButton onClick={toggleDropdownState}>
-                  {selectedCategory.name}
-                </S.CategoryButton>
-                {isShownDropdown && (
-                  <S.CategoryBox
-                    className={isClosing ? 'close' : ''}
-                    animationTime={dropdownAnimationTime}
-                  >
-                    {categories &&
-                      categories.map(item => (
-                        <S.CategoryList
-                          onClick={changeCategory(item)}
-                          key={item.id}
-                        >
-                          {item.name}
-                        </S.CategoryList>
-                      ))}
-                  </S.CategoryBox>
-                )}
-              </S.Label>
-              <S.Label>
-                최대 인원 수
-                <S.Input
-                  type="number"
-                  min="1"
-                  max="99"
-                  placeholder="99"
-                  value={capacity}
-                  onChange={changeCapacity}
-                />
-              </S.Label>
-              <S.Label>
-                기간
-                <S.SingleLineContainer>
-                  <S.Input
-                    type="date"
-                    value={startDate}
-                    onChange={setStartDate}
-                    min={getNewDateString('day')}
-                  />
-                  ~
-                  <S.Input
-                    type="date"
-                    value={endDate}
-                    onChange={setEndDate}
-                    min={startDate || getNewDateString('day')}
-                  />
-                </S.SingleLineContainer>
-              </S.Label>
-              <S.Label>
-                상세 일정
-                <CalendarEditor
-                  useScheduleState={useScheduleState}
-                  duration={duration}
-                />
-              </S.Label>
-              <S.Label>
-                모집 마감일
-                <S.Input
-                  type="datetime-local"
-                  value={deadline}
-                  onChange={setDeadline}
-                  min={getNewDateString('min')}
-                />
-              </S.Label>
-              <S.Label>
-                장소
-                <S.Input type="text" value={location} onChange={setLocation} />
-              </S.Label>
-              <S.Label>
-                설명
-                <S.TextArea
-                  value={description}
-                  onChange={setDescription}
-                  maxLength={GROUP_RULE.DESCRIPTION.MAX_LENGTH}
-                />
-              </S.Label>
-              <S.ButtonContainer>
-                <S.EditButton onClick={groupEdit}>수정</S.EditButton>
-                <S.RevertButton onClick={setOffModal}>취소하기</S.RevertButton>
-              </S.ButtonContainer>
-            </>
-          )}
-        </S.InputContainer>
+        <Inputs
+          useNameState={useNameState}
+          useSelectedCategoryState={useSelectedCategoryState}
+          useCapacityState={useCapacityState}
+          useDateState={useDateState}
+          useScheduleState={useScheduleState}
+          useDeadlineState={useDeadlineState}
+          useLocationState={useLocationState}
+          useDescriptionState={useDescriptionState}
+        />
+        <S.ButtonContainer>
+          <S.EditButton onClick={editGroup}>수정</S.EditButton>
+          <S.RevertButton onClick={setOffModal}>취소하기</S.RevertButton>
+        </S.ButtonContainer>
       </S.Container>
     </Modal>
   );
